🎨 style(README.md): fix formatting and add a new link The formatting of the README.md file has been improved by adding a space after the '#' symbol in the title. Additionally, a new link to a tutorial on how to install cloudflared on UnRAID has been added.
📝 docs(cloudflared.md): add instructions for setting up Cloudflare and UnRAID
This commit adds instructions for setting up Cloudflare and UnRAID. It includes steps for configuring a domain, setting up a tunnel, and creating a new docker network. It also includes screenshots to help guide the user through the process.

🎨 style(cloudflared.md): improve formatting and add headings
This commit improves the formatting of the cloudflared.md file and adds headings to make it easier to read and follow along. No functional changes were made.

🌅 chore(UnRAID): add UnRAID_Guacamole_port.png image
This commit adds a new image file, UnRAID_Guacamole_port.png, to the UnRAID/images directory. The image is used to illustrate the configuration of the Guacamole port on UnRAID.